Move `getPhoto` to a separate file to fix hashdev build
